@@ -8,7 +8,7 @@ import sys
 import os
 from datetime import timedelta
 from time import perf_counter
-from mimetypes import guess_type, init as mt_init, add_type
+from mimetypes import guess_type, init as mt_init
 from typing import Callable, Dict, List
 from xml.etree import ElementTree
 from pathlib import Path
@@ -20,7 +20,7 @@ import trio
 import requests
 import humanize
 
-from everyday_scripts.scriptlib import *
+from everyday_scripts.scriptlib import logging_init, sig_quit_clean
 
 DESCRIPTION = """Move files to different destinations based on their mime-type.
 
@@ -53,7 +53,7 @@ class Metrics:
         print("\nSTATS:")
         for attr in ["moves", "copies", "errors"]:
             print(f"  - {attr:5}\t: {cls.__dict__[attr]:5}")
-        print(f"  - types\t:")
+        print("  - types\t:")
         for k in sorted(cls.type_counts.keys()):
             print(f"  \t- {k}\t: {cls.type_counts[k]:5}")
 
@@ -202,17 +202,13 @@ def valid_dir_or_quit(path: str, name: str) -> str:
 
 
 def main():
-    parser = argparse.ArgumentParser(
-        "mmm", description=DESCRIPTION, formatter_class=argparse.RawDescriptionHelpFormatter
-    )
+    parser = argparse.ArgumentParser("mmm", description=DESCRIPTION, formatter_class=argparse.RawDescriptionHelpFormatter)
     parser.add_argument("--verbose", "-v", action="store_true", help="List actions")
     parser.add_argument("--debug", "-d", action="store_true", help="Debug level logging")
     parser.add_argument("--mt-download", "-m", action="store_true", help="Download mime types reference")
     parser.add_argument("--dry-run", "-n", action="store_true", help="Dry run")
     parser.add_argument("--copy", "-c", action="store_true", help="Copy, don't move")
-    parser.add_argument(
-        "--parallel", "-P", type=int, default=PARALLELISM, help="Number of parallel operations (default: %(default)d)"
-    )
+    parser.add_argument("--parallel", "-P", type=int, default=PARALLELISM, help="Number of parallel operations (default: %(default)d)")
     parser.add_argument("src", metavar="SOURCE_DIR", help="Source directory")
     parser.add_argument("dst", metavar="DST_PATTERN", nargs="+", help="Destination patterns in TYPE:DIRECTORY format")
     args = parser.parse_args()
